i18n example fix (#3482)
Today I was implementing an i18n according to this example in one of my projects. Unfortunately, current example fails if (for example) our `defaultLocale` is `en` and our url is `example.com/en/` or `example.com/en/news`: middleware redirects into `example.com//` and `example.con//news` respectively.

I've fixed the example (not sure if it is the most elegant way to do it though), so it replaces `/en/` with `/`.