fix(vue-app): Followup for fix for handle prefetch error (#5687)
After more investigation I've came to conclusion that keeping previous
behavior of marking component as prefetched before we know if it
loaded/failed, is the right thing to do. That's because otherwise, when
processing many nuxt-links within a page, that code will trigger multiple
times, with same component list and trigger loading of them multiple
times (which might be optimized by webpack code anyway but still...).