fix(vue-app): initial "synthetic" route invalid when router in hash mode
On the client, Nuxt creates "synthetic" route initially, before VueRouter
has initialized so that plugins and middlewares have something to work
with at that early stage already. The problem was that initial route
was created without taking router's `mode` into consideration so it
created route with normal path rather than hash-based path. That confused
nuxt-i18n's logic that tries to redirect user to correct locale on
initial navigation.

As for tests:
 - created new `spa-hash` fixture config so that we can run tests in hash mode
 - new fixture is using files from `spa` fixture to avoid duplication
 - change `spa.test.js` testsuite to run same tests on both fixtures
 - disable `/тест雨 (test non ascii route)` test in hash mode as it is failing
   for unrelated reason (I haven't investigated deeper).
 - add `Initial route has correct fullPath` test to test the bug fixed here

Resolves https://github.com/nuxt-community/nuxt-i18n/issues/523