Fix bug where `redirect('...')` strips trailing slash from URL
When calling `redirect('...')` with an external (non-relative) URL,
and when that URL has a trailing slash, Nuxt will remove the trailing
slash before performing the redirect. This breaks when the external URL
requires that trailing slash!

The cause of this is simple. Nuxt was filtering empty components
form the URL's path. This is over-paranoid and (in this case) was
causing a bug where the trailing slash was being removed (because
there was an empty component path after that trailing slash, which
was being filtered out before re-joining the path back together.

The fix is simple. Don't filter empty components from the path.
This type of paranoid URL cleaning should not be the responsibility
of the `redirect()` function.