fix: serializeFunction helper fails with certain functions
A function shorthand with certain statements inside failed to
serialize properly using serializeFunction.

The issue was due to the replacement patterns not considering there
being a space between the function name and the parenthesis before the
arguments.

Changed the simple string pattern to regexp to be able to match any
number of whitespace characters between those.

Also changed the test file to not be transformed by babel-jest as that
has affected the input passed to "serializeFunction" and made some tests
invalid. For example, the new test didn't even fail with babel-jest as
it was transforming the test input in a way that made it pass.

Resolves #8779