docs(StateManagement): Update 7.state-management.md
The website docs say not to use the practice of ref() outside of a component's setup. It's not clear what's an example of a bad practice of using it. Adding an [example shared by Daniel Roe](https://x.com/danielcroe/status/1808232524956455106) to clarify one use-case (that shouldn't be followed).

This update to the README help clarifies it.