fix(nuxt): ensure getCachedData function is returning cache if possible
Unless there is manual refresh or refresh() has been explicitly called,
return the cache if available. use payload data first, fallback to
static if that is not available. Returns undefined when no cache has
been found, triggering a fetch. Previous default function only returned
cache when hydrating, which is not desired in most scenarios, therefore
the default is changed.