Update 07.routing.md
The version 3.18 Getting Started->07.Routing with subsection Navigation references adding the navigation to pages/app.vue at the current stage of the documentation, read sequentially. This should be a mistake. What is probably meant is the pages/index.vue. 

This is further illustrated as in the Docs for further reading, it is correctly labeled as index.vue, as opposed to giving newcomers following along the wrong idea that you should move app.vue into pages and add the navigation elements there. This would not work either, as then there are no <NuxtPage> in the app/app.vue serving as the entrypoint for the pages API